Add CitationFinder class
