+++ src/main/java/org/jabref/gui/entryeditor/LatexCitationsTabViewModel.java
@@ -1,13 +1,8 @@
 package org.jabref.gui.entryeditor;
 
-import java.io.IOException;
-import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.Collection;
-import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.Future;
-import java.util.stream.Stream;
 
 import javafx.beans.property.ObjectProperty;
 import javafx.beans.property.ReadOnlyListWrapper;
@@ -23,12 +18,11 @@ import org.jabref.gui.util.BackgroundTask;
 import org.jabref.gui.util.DirectoryDialogConfiguration;
 import org.jabref.gui.util.TaskExecutor;
 import org.jabref.logic.l10n.Localization;
-import org.jabref.logic.texparser.DefaultLatexParser;
+import org.jabref.logic.texparser.CitationFinder;
 import org.jabref.logic.util.io.FileUtil;
 import org.jabref.model.database.BibDatabaseContext;
 import org.jabref.model.entry.BibEntry;
 import org.jabref.model.texparser.Citation;
-import org.jabref.model.texparser.LatexParserResults;
 import org.jabref.preferences.PreferencesService;
 
 import org.slf4j.Logger;
@@ -50,11 +44,11 @@ public class LatexCitationsTabViewModel extends AbstractViewModel {
     private final TaskExecutor taskExecutor;
     private final DialogService dialogService;
     private final ObjectProperty<Path> directory;
+    private final CitationFinder citationFinder;
     private final ObservableList<Citation> citationList;
     private final ObjectProperty<Status> status;
     private final StringProperty searchError;
     private Future<?> searchTask;
-    private LatexParserResults latexParserResults;
     private BibEntry currentEntry;
 
     public LatexCitationsTabViewModel(BibDatabaseContext databaseContext,
@@ -67,6 +61,7 @@ public class LatexCitationsTabViewModel extends AbstractViewModel {
         this.dialogService = dialogService;
         this.directory = new SimpleObjectProperty<>(databaseContext.getMetaData().getLatexFileDirectory(preferencesService.getFilePreferences().getUserAndHost())
                                                                    .orElse(FileUtil.getInitialDirectory(databaseContext, preferencesService.getFilePreferences().getWorkingDirectory())));
+        this.citationFinder = new CitationFinder(directory.get());
         this.citationList = FXCollections.observableArrayList();
         this.status = new SimpleObjectProperty<>(Status.IN_PROGRESS);
         this.searchError = new SimpleStringProperty("");
@@ -103,7 +98,10 @@ public class LatexCitationsTabViewModel extends AbstractViewModel {
     }
 
     private void startSearch(String citeKey) {
-        searchTask = BackgroundTask.wrap(() -> searchAndParse(citeKey))
+        // we need to check whether the user meanwhile set the LaTeX file directory or the database changed locations
+        checkDirectory();
+
+        searchTask = BackgroundTask.wrap(() -> citationFinder.searchAndParse(citeKey))
                                    .onRunning(() -> status.set(Status.IN_PROGRESS))
                                    .onSuccess(result -> {
                                        citationList.setAll(result);
@@ -116,6 +114,16 @@ public class LatexCitationsTabViewModel extends AbstractViewModel {
                                    .executeWith(taskExecutor);
     }
 
+    public void checkDirectory() {
+        Path newDirectory = databaseContext.getMetaData().getLatexFileDirectory(preferencesService.getFilePreferences().getUserAndHost())
+                                           .orElse(FileUtil.getInitialDirectory(databaseContext, preferencesService.getFilePreferences().getWorkingDirectory()));
+
+        if (!newDirectory.equals(directory.get())) {
+            directory.set(newDirectory);
+            citationFinder.setDirectory(newDirectory);
+        }
+    }
+
     private void cancelSearch() {
         if (searchTask == null || searchTask.isCancelled() || searchTask.isDone()) {
             return;
@@ -123,42 +131,6 @@ public class LatexCitationsTabViewModel extends AbstractViewModel {
 
         status.set(Status.IN_PROGRESS);
         searchTask.cancel(true);
-    }
-
-    private Collection<Citation> searchAndParse(String citeKey) throws IOException {
-        // we need to check whether the user meanwhile set the LaTeX file directory or the database changed locations
-        Path newDirectory = databaseContext.getMetaData().getLatexFileDirectory(preferencesService.getFilePreferences().getUserAndHost())
-                                           .orElse(FileUtil.getInitialDirectory(databaseContext, preferencesService.getFilePreferences().getWorkingDirectory()));
-
-        if (latexParserResults == null || !newDirectory.equals(directory.get())) {
-            directory.set(newDirectory);
-
-            if (!newDirectory.toFile().exists()) {
-                throw new IOException("Current search directory does not exist: %s".formatted(newDirectory));
-            }
-
-            List<Path> texFiles = searchDirectory(newDirectory);
-            LOGGER.debug("Found tex files: {}", texFiles);
-            latexParserResults = new DefaultLatexParser().parse(texFiles);
-        }
-
-        return latexParserResults.getCitationsByKey(citeKey);
-    }
-
-    /**
-     * @param directory the directory to search for. It is recursively searched.
-     */
-    private List<Path> searchDirectory(Path directory) {
-        LOGGER.debug("Searching directory {}", directory);
-        try (Stream<Path> paths = Files.walk(directory)) {
-            return paths
-                    .filter(Files::isRegularFile)
-                    .filter(path -> path.toString().endsWith(TEX_EXT))
-                    .toList();
-        } catch (IOException e) {
-            LOGGER.error("Error while searching files", e);
-            return List.of();
-        }
     }
 
     public void setLatexDirectory() {
@@ -172,7 +144,6 @@ public class LatexCitationsTabViewModel extends AbstractViewModel {
     }
 
     public void refreshLatexDirectory() {
-        latexParserResults = null;
         init(currentEntry);
     }
